refactor(human resources): refactor the HR module and generate student's swagger file
refactor the HR components into a single module, generate the swagger file of the student module
