+++ src/app.ts
@@ -1,13 +1,11 @@
 import express from 'express';
 import cors from 'cors';
 import teacherRoute from './routes/teacher.route';
-import studentRoute from './routes/student.routes';
-import leaveRoute from './routes/leaves.route';
-import scheduleRoute from './routes/schedule.route';
-import paySleepRoute from './routes/paysleep.route';
+import humanResourcesRoute from './routes/humanResources.route';
 import swaggerUi from 'swagger-ui-express';
 import swaggerDocs from '../swagger';
 import { PrismaClient } from '@prisma/client';
+import studentRoute from './routes/student.route';
 
 // create new prisma object
 
@@ -23,11 +21,8 @@ app.use(cors());
 
 // main route
 app.use('/api/teacher', teacherRoute);
+app.use('/api/hr', humanResourcesRoute);
 app.use('/api/student', studentRoute);
-
-app.use('/api/leave', leaveRoute);
-app.use('/api/paysleep', paySleepRoute);
-app.use('/api/schedule', scheduleRoute);
 // Set up Swagger UI
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocs));
