--- src/routes/student.routes.ts
@@ -1,17 +1,0 @@
-import { Router } from 'express';
-import { authMiddleware } from '../middlewares/auth.middleware';
-import {
-  createStudent,
-  deleteStudent,
-  getStudent,
-  updateStudent,
-} from '../controllers/student.controller';
-
-const router = Router();
-
-router.post('/create', createStudent);
-router.put('/update', authMiddleware, updateStudent);
-router.delete('/delete', authMiddleware, deleteStudent);
-router.get('/all-students', getStudent);
-
-export default router;
